refactor: remove eda/utils,  rename folders: eda -> analysis,  jiratoolkit -> jira_analysis
